added user article association

+++ api/controllers/ArticleController.js
@@ -10,7 +10,7 @@ module.exports = {
     // Index
     index : function(req, res) {
         console.log('GET - /articles');
-         return Article.find({}).exec(function(err, articles) {
+         return Article.find({ owner: req.user[0].id }).populate('owner').exec(function(err, articles) {
             return res.view('articles/index.ejs', { articles: articles });
         });
     },
@@ -18,7 +18,7 @@ module.exports = {
     // Show action
     show : function(req, res) {
         console.log('GET - /article/:id');
-        return Article.findOne({ id: req.params.id }).exec(function(err, article) {
+        return Article.findOne({ id: req.params.id, owner: req.user[0].id }).exec(function(err, article) {
             if(!article){
                 res.statusCode = 404;
                 return res.send({ error: 'Article not found.' });
@@ -46,7 +46,8 @@ module.exports = {
         console.log('POST - /article');
         var articleObj = {
                             title   : req.body.title, 
-                            content : req.body.content
+                            content : req.body.content,
+                            owner   : req.user[0].id
         };
 
         return Article.create(articleObj).exec(function(err, article){
@@ -64,7 +65,7 @@ module.exports = {
     // Edit action
     edit : function(req, res) {
         console.log('GET - /article/:id/edit');
-        return Article.findOne({ id: req.params.id }).exec(function(err, article){
+        return Article.findOne({ id: req.params.id , owner: req.user[0].id }).exec(function(err, article){
              if(!article){
                 res.statusCode = 404;
                 return res.send({ error: 'Article not found.' });
@@ -86,7 +87,8 @@ module.exports = {
         console.log('PUT - /article/:id');
         var articleObj = {
                             title   : req.body.title,
-                            content : req.body.content
+                            content : req.body.content,
+                            owner   : req.user[0].id
         }
         return Article.update({ id: req.params.id }, articleObj).exec(function(err, article){
             if(!err){
@@ -103,7 +105,7 @@ module.exports = {
     // Destroy action
     destroy: function(req, res) {
         console.log('DELETE - /article:id');
-        return Article.destroy({ id: req.params.id }).exec(function(err){
+        return Article.destroy({ id: req.params.id, owner: req.user[0].id }).exec(function(err){
             if(err){
                 res.statusCode = 500;
                 console.log('Internal server error(%d): %s', res.statusCode, err)
